requirements/base.txt
aiohttp==3.8.5
aiomqtt==1.2.1
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.2.0
charset-normalizer==3.3.2
Django==5.0.1
django-cors-headers==4.3.1
django-debug-toolbar==4.3.0
django-environ==0.9.0
django-extensions==3.2.3
djangorestframework==3.14.0
djangorestframework-simplejwt==5.2.2
drf-spectacular==0.24.2
environ==1.0
frozenlist==1.4.1
idna==3.6
inflection==0.5.1
jsonschema==4.20.0
jsonschema-specifications==2023.12.1
multidict==6.0.4
pillow==10.2.0
psycopg==3.1.17
psycopg2-binary==2.9.9
psycopg2-pool==1.1
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
referencing==0.32.1
rest-framework-simplejwt==0.0.2
rpds-py==0.16.2
sqlparse==0.4.4
typing_extensions==4.9.0
uritemplate==4.1.1
yarl==1.9.4
